Add files via upload
DMART Sales prediction is a type of problem in the field of machine learning that involves using historical sales data to predict future sales for a retail store such as DMART. The goal is to create a model that can accurately predict the sales of various products in different outlets of the Big Mart store based on various factors such as price, marketing campaigns, competition, etc.

To solve this problem, a supervised machine learning algorithm is typically used, where the model is trained on the historical sales data and uses this information to make predictions on new, unseen data. The process involves selecting relevant features, cleaning and preprocessing the data, splitting the data into training and testing sets, training the model, and finally, evaluating the model's performance on the test data. Common algorithms used for this type of problem include linear regression, decision trees, and random forests.

The accuracy of the sales prediction model is crucial as it can help the store make informed decisions on various aspects such as product stocking, pricing strategies, and marketing campaigns, leading to increased sales and profits.